moved nft components to folder,  fixed selection issues
