--- packages/bridge-ui-v2/src/components/NFTList/index.ts
@@ -1,1 +1,0 @@
-export { default as NFTList } from './NFTList.svelte';
