+++ packages/bridge-ui-v2/src/components/NFTs/NFTDisplay.svelte
@@ -1,0 +1,31 @@
+<script lang="ts">
+  import { t } from 'svelte-i18n';
+
+  import { LoadingMask } from '$components/LoadingMask';
+  import type { NFT } from '$libs/token';
+  import { network } from '$stores/network';
+
+  import { NFTCardGrid } from './NFTCards';
+  import { NFTList } from './NFTList';
+  import { NFTView } from './types';
+
+  export let loading: boolean;
+
+  export let viewOnly = false;
+
+  export let nfts: NFT[] | null = [];
+
+  export let nftView: NFTView = NFTView.LIST;
+</script>
+
+<div class="relative max-h-[350px] min-h-[350px] bg-neutral rounded-[20px] overflow-hidden">
+  <div class="max-h-[350px] min-h-[350px] overflow-y-auto py-2 px-[20px]">
+    {#if loading}
+      <LoadingMask spinnerClass="border-white" text={$t('messages.bridge.nft_scanning')} />
+    {:else if nftView === NFTView.LIST && nfts}
+      <NFTList bind:nfts chainId={$network?.id} {viewOnly} />
+    {:else if nftView === NFTView.CARDS && nfts}
+      <NFTCardGrid bind:nfts {viewOnly} />
+    {/if}
+  </div>
+</div>
