+++ packages/bridge-ui-v2/src/components/NFTs/NFTInfoDialog.svelte
