+++ packages/bridge-ui-v2/src/components/NFTs/NFTList/NFTList.svelte
