+++ packages/bridge-ui-v2/src/components/NFTs/NFTList/index.ts
@@ -1,0 +1,2 @@
+export { default as NFTList } from './NFTList.svelte';
+export { default as NFTListItem } from './NFTListItem.svelte';
