+++ packages/bridge-ui-v2/src/components/NFTs/index.ts
@@ -1,0 +1,1 @@
+export { default as NFTDisplay } from './NFTDisplay.svelte';
