+++ packages/bridge-ui-v2/src/components/NFTs/types.ts
@@ -1,0 +1,4 @@
+export enum NFTView {
+  CARDS,
+  LIST,
+}
